+++ cli/run/getConfigPath.ts
@@ -1,14 +1,14 @@
-import { readdirSync, realpathSync } from 'fs';
+import { readdirSync } from 'fs';
 import * as path from 'path';
 import relative from 'require-relative';
 import { handleError } from '../logging';
 
 const DEFAULT_CONFIG_BASE = 'rollup.config';
 
-export function getConfigPath(commandConfig: any): string {
+export function getConfigPath(commandConfig: string | true): string {
 	const cwd = process.cwd();
 	if (commandConfig === true) {
-		return path.join(cwd, findConfigFileNameInCwd());
+		return path.resolve(findConfigFileNameInCwd());
 	}
 	if (commandConfig.slice(0, 5) === 'node:') {
 		const pkgName = commandConfig.slice(5);
@@ -28,7 +28,7 @@ export function getConfigPath(commandConfig: any): string {
 			}
 		}
 	}
-	return realpathSync(commandConfig);
+	return path.resolve(commandConfig);
 }
 
 function findConfigFileNameInCwd(): string {