+++ test/cli/samples/symlinked-named-config/my.rollup.config.js
@@ -1,0 +1,1 @@
+rollup.config.symlinked